include flexslider as a normal JSFooter File
Lets not include the flexslider file as a Lib. It could lead to conflict with the order of integrated Libs (for example, jQuery as a FooterLib being included AFTER the flexslider file, which could result in a jQuery not defined error). Or would you define the flexslider plugin as a LIB?