Fix: Not set gets lost after second save
Well TYPO3 seems to handle an empty string as not set, leading to tca to auto-select the first select option after save.

As a little hack I simple defined '---' as not-set value. Which ofc causes a class '---' in frontend, but i think for now it will not do harm anywhere as it is very unlikely to be styled by anyone that uses this extension. I might be wrong in the future and recovery is definitly possible if someone steps on this as a bug.